Launcher: move http related methods to HTTP module
We need to get these methods out of Launcher.
