+++ calabash-cucumber/lib/calabash-cucumber.rb
@@ -4,6 +4,7 @@ require "calabash-cucumber/dot_dir"
 require "calabash-cucumber/store/preferences"
 require "calabash-cucumber/usage_tracker.rb"
 require "calabash-cucumber/dylibs"
+require "calabash-cucumber/http/http"
 require 'calabash-cucumber/core'
 require 'calabash-cucumber/tests_helpers'
 require 'calabash-cucumber/keyboard_helpers'
